+++ actix-web/CHANGES.md
@@ -5,11 +5,20 @@
 - Add `ContentDisposition::attachment` constructor. [#2867]
 - Add `ErrorHandlers::default_handler()` (as well as `default_handler_{server, client}()`) to make registering handlers for groups of response statuses easier. [#2784]
 - Add `Logger::custom_response_replace()`. [#2631]
+- Add rudimentary redirection service at `web::redirect()` / `web::Redirect`. [#1961]
+- Add `guard::Acceptable` for matching against `Accept` header mime types. [#2265]
+- Add fallible versions of test helpers: `try_call_service`, `try_call_and_read_body_json`, `try_read_body`, and `try_read_body_json`. [#2961]
 
+### Fixed
+- Add `Allow` header to `Resource`'s default responses when no routes are matched. [#2949]
+
+[#1961]: https://github.com/actix/actix-web/pull/1961
+[#2265]: https://github.com/actix/actix-web/pull/2265
 [#2631]: https://github.com/actix/actix-web/pull/2631
 [#2784]: https://github.com/actix/actix-web/pull/2784
 [#2867]: https://github.com/actix/actix-web/pull/2867
-
+[#2949]: https://github.com/actix/actix-web/pull/2949
+[#2961]: https://github.com/actix/actix-web/pull/2961
 
 ## 4.2.1 - 2022-09-12
 ### Fixed